Version 1.2: Realtime database with read, write, update and delete. Usability and layout changed. Added a progressbar for both features and cleaned the code and also regrouped all bmi and bfp stuff together( categorisation, result messages, constants .. all in MainActivity).
A User can calculate his BMI & BFP, the User need's to introduce his weight, height, age, sex , then he chose witch calculation he wants and he can chose to Store his information (with results in SQLite db) or not.
Once the User introduced this variables they will be stored depending of user preferences.
(If the user chose to store the info => he can close the app and come back later -> the info will be filled in the app)
**
In this version I created a SQLite db => you can see it in The Device File Explorer => data > basicbfpapplication => database.
(You can also read the file in a DB browser.)
Here I used the class SQLiteOpenHelper and a Switch to chose if the info will be stored or not.